Begin refactoring of sp_PE_JobMatrix to ready it for production use
The sp_PE_JobMatrix proc is very rough. This commit implements the
first set of changes. Big blocks of logic have been moved to helper
procedures in the CorePE schema. The job run simulation logic has
been refactored and is simpler and more accurate. However, this
commit does not represent a production-ready version of the JobMatrix
procedure, and further commits will be forthcoming on this branch.