Remove predictive logic, add Queries logic, rework Help section
The Predictive Matrix has proven to be too difficult due to the lack of
MSFT support in finding the next job start time for a schedule or
understanding the way that "Every X Weeks/Months" logic is calculated.
All logic related to the predictive matrix has been removed (files with
the logic have been saved to the Stashed folder). The Queries
functionality has been implemented. The Help code and contents now matches
the other sp_ procs in PerformanceEye and accurately reflects the
functionality of the current code. Thus, sp_PE_JobMatrix is ready for
final testing before release.